Update README
Remove example of using a non-static method to modify a player class.
This is often not the cleanest way to do it, especially if there are
multiple instances of the game object. Instead a command
can simply receive a unique identifier as its first argument, and
use that to look up a reference to the game object.