Keep .meta files in version control, ignore in root dir only
Not having meta files causes them to be regenerated on every user's
computer, each with a different GUID. This makes it impossible to open
the scene/prefab files on different computers - Terminal will instead be
represented as "missing script".

This change fixes this problem by only ignoring meta files in root dir -
meta files in CommandTerminal and all other subdirectories will not be
ignored by git.